# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TimeTracking",
    "description": "TimeTracking API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3002/"
    }
  ],
  "parameters": [
    {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": { "type": "integer", "default": 1 }
    },
    {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": { "type": "integer", "default": 10 }
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "description": "Créer un nouvel utilisateur",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "user@example.com" },
                  "username": { "type": "string", "example": "user123" },
                  "password": { "type": "string", "example": "password123" }
                },
                "required": ["email", "username", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Utilisateur créé avec succès" },
          "400": { "description": "Erreur de validation" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "description": "Connexion d'un utilisateur",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string", "example": "user123" },
                  "password": { "type": "string", "example": "password123" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Connexion réussie" },
          "401": { "description": "Identifiants incorrects" }
        }
      }
    },
    "/projects": {
      "get": {
        "tags": ["Projects"],
        "summary": "Lister tous les projets de l'utilisateur",
        "responses": {
          "200": { "description": "Liste des projets" }
        }
      },
      "post": {
        "tags": ["Projects"],
        "summary": "Créer un nouveau projet",
        "responses": {
          "201": { "description": "Projet créé avec succès" },
          "400": { "description": "Erreur de validation" }
        }
      }
    },
    "/projects/{slug}": {
      "get": {
        "tags": ["Projects"],
        "summary": "Obtenir un projet par son slug",
        "parameters": [{ "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Projet trouvé" },
          "404": { "description": "Projet non trouvé" }
        }
      },
      "put": {
        "tags": ["Projects"],
        "summary": "Mettre à jour un projet",
        "parameters": [{ "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Projet mis à jour" },
          "404": { "description": "Projet non trouvé" }
        }
      },
      "delete": {
        "tags": ["Projects"],
        "summary": "Supprimer un projet",
        "parameters": [{ "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Projet supprimé" },
          "404": { "description": "Projet non trouvé" }
        }
      }
    },
    "/projects/{slug}/tasks": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Obtenir les tâches d'un projet",
        "parameters": [{ "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Liste des tâches" },
          "404": { "description": "Projet non trouvé" }
        }
      },
      "post": {
        "tags": ["Tasks"],
        "summary": "Créer une tâche dans un projet",
        "parameters": [{ "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "201": { "description": "Tâche créée" },
          "400": { "description": "Erreur de validation" }
        }
      }
    },
    "/projects/{slug}/tasks/{taskSlug}": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Obtenir une tâche spécifique",
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "taskSlug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Tâche trouvée" },
          "404": { "description": "Tâche non trouvée" }
        }
      },
      "put": {
        "tags": ["Tasks"],
        "summary": "Mettre à jour une tâche",
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "taskSlug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Tâche mise à jour" },
          "404": { "description": "Tâche non trouvée" }
        }
      },
      "delete": {
        "tags": ["Tasks"],
        "summary": "Supprimer une tâche",
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "taskSlug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Tâche supprimée" },
          "404": { "description": "Tâche non trouvée" }
        }
      }
    }
  }
}